set cap for estimated obstacle size

+++ include/onboard_detector/utils.h
@@ -17,11 +17,11 @@ namespace onboardDetector{
     struct box3D
     {
         /* data */
-        float x, y, z;
-        float x_width, y_width, z_width;
-        float id;
-        float Vx, Vy, Vz;
-        float Ax, Ay, Az;
+        double x, y, z;
+        double x_width, y_width, z_width;
+        double id;
+        double Vx, Vy, Vz;
+        double Ax, Ay, Az;
         bool is_human=false; // false: not detected by yolo as dynamic, true: detected by yolo
         bool is_dynamic=false; // false: not detected as dynamic(either yolo or classificationCB), true: detected as dynamic
         bool fix_size=false; // flag to force future boxes to fix size
